Refactor out common usb_serial code
